# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/reviews/{reviewId}/{productId}:
    put:
      tags:
      - review-rest-controller
      operationId: replaceReview
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewReviewRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
    delete:
      tags:
      - review-rest-controller
      operationId: deleteReview
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
...
  /api/v1/user:
    post:
      tags:
      - register-rest-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
  /api/v1/user/image:
    post:
      tags:
      - register-rest-controller
      operationId: createUserImage
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
              required:
              - imageFile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/reviews/{productId}:
    post:
      tags:
      - review-rest-controller
      operationId: createReview
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewReviewRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
...
  /api/v1/reviews/{reviewId}:
    patch:
      tags:
      - review-rest-controller
      operationId: reportReview
      parameters:
      - name: userId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
...
  /api/v1/reviews/{id}:
    get:
      tags:
      - review-rest-controller
      operationId: getReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
...
components:
  schemas:
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        address:
          type: string
        banned:
          type: boolean
        reports:
          type: integer
          format: int32
        imageUrl:
          type: string
        roles:
          type: array
          items:
            type: string
    NewReviewRequestDTO:
      type: object
      properties:
        rating:
          type: integer
          format: int32
        reviewText:
          type: string
    ReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        user:
          $ref: "#/components/schemas/UserDTO"
        rating:
          type: integer
          format: int32
        reviewText:
          type: string
        reported:
          type: boolean
        own:
          type: boolean
        ratingStars:
          type: array
          items:
            type: boolean
        emptyStars:
          type: array
          items:
            type: boolean
...
    NewUserDTO:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        encodedPassword:
          type: string
        confirmPassword:
          type: string
...